authentication & authorization implemented

+++ src/services/notes/controller/note-controller.js
@@ -1,13 +1,19 @@
 import response from "../../../utils/response.js";
-import NoteRepositories from "../repositories/note-repositories.js";
-import { InvariantError, NotFoundError } from "../../../exceptions/index.js";
+import NoteRepositories from "../repositories/note-repository.js";
+import {
+  InvariantError,
+  NotFoundError,
+  AuthorizationError,
+} from "../../../exceptions/index.js";
 
 export const createNote = async (req, res, next) => {
   const { title, tags, body } = req.validated;
+  const { id: owner } = req.user;
   const note = await NoteRepositories.createNote({
     title,
     tags,
     body,
+    owner,
   });
 
   if (!note) {
@@ -17,12 +23,22 @@ export const createNote = async (req, res, next) => {
 };
 
 export const getNotes = async (req, res) => {
-  const notes = await NoteRepositories.getNotes();
-  return response(res, 200, "success", { notes });
+  const { id: owner } = req.user;
+  const notes = await NoteRepositories.getNotes(owner);
+  return response(res, 200, "Catatan sukses ditampilkan", { notes });
 };
 
 export const getNoteById = async (req, res, next) => {
   const { id } = req.params;
+  const { id: owner } = req.user;
+  const isOwner = await NoteRepositories.verifyNoteOwner(id, owner);
+
+  if (!isOwner) {
+    return next(
+      new AuthorizationError("Anda tidak berhak mengakses resource ini"),
+    );
+  }
+
   const note = await NoteRepositories.getNoteById(id);
 
   if (!note) {
@@ -34,6 +50,16 @@ export const getNoteById = async (req, res, next) => {
 export const updateNoteById = async (req, res, next) => {
   const { id } = req.params;
   const { title, body, tags } = req.validated;
+  const { id: owner } = req.user;
+
+  const isOwner = await NoteRepositories.verifyNoteOwner(id, owner);
+
+  if (!isOwner) {
+    return next(
+      new AuthorizationError("Anda tidak berhak mengakses resource ini"),
+    );
+  }
+
   const note = await NoteRepositories.editNote({ id, title, body, tags });
 
   if (!note) {
@@ -42,9 +68,19 @@ export const updateNoteById = async (req, res, next) => {
   return response(res, 200, "Catatan berhasil diperbarui", { note });
 };
 
-export const deleteNoteById = (req, res, next) => {
+export const deleteNoteById = async (req, res, next) => {
   const { id } = req.params;
-  const note = NoteRepositories.deleteNote(id);
+  const { id: owner } = req.user;
+
+  const isOwner = await NoteRepositories.verifyNoteOwner(id, owner);
+
+  if (!isOwner) {
+    return next(
+      new AuthorizationError("Anda tidak berhak mengakses resource ini"),
+    );
+  }
+
+  const note = await NoteRepositories.deleteNote(id);
 
   if (!note) {
     return next(new NotFoundError("Catatan tidak ditemukan"));
